Change all 'int' in for loops to 'size_t' to eliminate signed/unsigned comparison warnings

+++ Competition.cpp
@@ -6,7 +6,7 @@ int Competition::countRemainingCities(char letter)
 {
     int count = 0;
 
-    for (int i = 0; i < all_cities.size(); ++i)
+    for (size_t i = 0; i < all_cities.size(); ++i)
     {
         if (all_cities[i][0] == letter && checkCity(all_cities[i]))
         {
@@ -22,7 +22,7 @@ string Competition::getCity(char start_letter)
     // Get all valid cities with the required letter
     vector<string> valid_cities;
 
-    for (int i = 0; i < all_cities.size(); i++)
+    for (size_t i = 0; i < all_cities.size(); i++)
     {
         if (all_cities[i][0] == start_letter && checkCity(all_cities[i]))
         {
@@ -39,7 +39,7 @@ string Competition::getCity(char start_letter)
     string best_city = valid_cities.front();
     int options = countRemainingCities(valid_cities[0].back());
 
-    for (int i = 1; i < valid_cities.size(); i++)
+    for (size_t i = 1; i < valid_cities.size(); i++)
     {
         char last_letter = valid_cities[i].back();
         int remaining_options = countRemainingCities(last_letter);
